Solution to Check if Temperature is Freezing

+++ Check if Temperature is Freezing@@6788dd63490637cfc44e1ba3/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(){
-    int temp:
-    scanf("%d",&temp);
+    float temp:
+    scanf("%f",&temp);
     if(temp>0){
         printf("Above Freezing");
     }
